requirements.txt
# EynatFormFunction/requirements.txt

azure-functions
fastapi
uvicorn
pydantic
pymssql